requirements.txt
asgiref==3.6.0
Django==4.1.7
django-stdimage==6.0.1
Pillow==9.5.0
sqlparse==0.4.3
